Split generics between domain & application
