+++ packster-core/src/operation/deploy.rs
@@ -1,6 +1,11 @@
 use std::path::{Path, PathBuf};
 
-use crate::{port::{FileSystem, Archiver}, domain::{Package, Deployment, DeployLocation, Checksum}, Result, path::Absolute};
+use crate::{
+    port::{FileSystem, Archiver},
+    domain::{Package, Deployment, DeployLocation, Checksum},
+    Result,
+    path::Absolute
+};
 
 use super::{
     Operation,
